requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
beautifulsoup4==4.7.1
brunns-matchers==2.1.1
brunns-row==2.0.0
certifi==2019.6.16
chardet==3.0.4
furl==2.0.0
idna==2.8
importlib-metadata==0.18
mbtest==2.2.0
more-itertools==7.1.0
orderedmultidict==1.0.1
packaging==19.0
pendulum==2.0.5
pluggy==0.12.0
py==1.8.0
pydantic==0.30
PyHamcrest==1.9.0
pyparsing==2.4.0
pytest==4.6.4
python-dateutil==2.8.0
pytzdata==2019.2
requests==2.22.0
six==1.12.0
soupsieve==1.9.2
urllib3==1.25.3
virtualenv==16.6.1
wcwidth==0.1.7
zipp==0.5.2
